upnp: move UpnpInstanceWrapper to upnp-wrapper
Move out UpnpInstanceWrapper into its own file
since it no longer has any dependencies on existing
modules and allow other modules to get references to it.
Move as well as the various helper functions that are
directly required by the UpnpInstanceWrapper class
and inline them for a small performance gain.

Signed-off-by: Hugo Beauzée-Luyssen <hugo@beauzee.fr>