contrib: Update SRT to 1.3.1
Almost the same as the previous one but squashing two patches which should have been and improving the inet_ntop() removing patch.

Skipping 1.3.0 as it was a bad version. Also removed all
custom patches are they don't seem to be required anymore, but
added new ones to make it build with MingW.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>