configure: Don't force _WIN32_WINNT to 0x0601 if it is set to a higher version
Since e807c8e91, we're explicitly unconditionally defining _WIN32_WINNT,
regardless if it was set before or not.

If it was set before (e.g. on the command line), to a higher version,
we should honor that.

This is cruicial for WinRT building, where _WIN32_WINNT is set to Windows
8 or higher on the command line.

EDIT: Adapted to a different default version in the 3.0 branch.

(cherry picked from commit 255e2ce27954004b8e6aa65ed37266c422f08a58)