cc: Only reset line position if rollup mode changed
The EIA-608 parser was always resetting the line position whenever
it received a rollup code.  However this would cause the text to
be rendered improperly if the rollup code was received in the
middle of an existing line and the mode was unchanged.

Only reset the line position if the rollup mode changes from the
existing value.

This problem was exhibited in streams from the "Steve Harvey show"
where whatever EIA-608 encoder they are using would sometimes send
rollup codes mid-row (causing the leading characters on that line to
be overwritten by whatever characters followed the rollup code).

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>
Signed-off-by: Francois Cartegnie <fcvlcdev@free.fr>