input: remove the last "*-es" var dependency
It was used to get the es id added by an input slave in order to enable it.

There is no real clean way to enable a future es_out_id from a future input
slave. The dumb way to do it is to listen to ES events and recover the last id
(what this commit is doing).