sout: chromecast: add stream info helpers
The previous implementation had hard-coded paths in several places,
and each place relied on other relevant part using the same paths.
From a maintainability perspective this is a mess, and bound to fail
one way or another.

These changes introduce helper-functions that can be called in places
where the values are needed, effectively adding a layer of protection
to maintainability issues by making sure everything is using the same
values.

Signed-off-by: Thomas Guillem <thomas@gllm.fr>