sout: chromecast: use stream info helpers in intf_sys_t
These changes includes usage of the relevant stream info helpers.

It also adds an additional slash to the art URI, so that items can be
accessed through "/chromecast/unique/unique/art/$ID", where $ID
denotes the art-id as used in the implementation, instead of
"[...]/art$ID". This is not a change in behavior as the URIs passed to
the device is per session, and the device is given notice of the URI
by us, it does not probe it implicitly.

Signed-off-by: Thomas Guillem <thomas@gllm.fr>
(cherry picked from commit d0ed421fb2f6760147dc29a814b661aadcc3dc2c)
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>