stream_extractor: archive: add archive_extractor_reset
In order for us to start over, such as a failed attempt of
dumb-seeking, we must reset the state of our extractor as if we just
opened the file.

This helper accomplishes that while also making sure that all
associated variables are back in a proper state.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>
(cherry picked from commit 6062e012d835e384cbfe8d1ae6ba2e2f17704815)
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>