core: add vlc_MetadataRequest
This function is to be used by entities within the core that would
like to spawn a metadata request. Such request will not have the
implicit recursiveness of libvlc_MetadataRequest (which is to be
called by users of libvlc).

Signed-off-by: Thomas Guillem <thomas@gllm.fr>
(cherry picked from commit d2c29640ffcc0af46e9abde972a10d636f688a29)
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>