vlc_input_item: rename INPUT_DURATION_INVALID to INPUT_DURATION_UNSET
A duration cannot be invalid and the define is used for items where a duration
should not be set like nodes and directories.

It's also used by input_item_New() on items that don't have the duration when
the source is parsed by may get it when parsed in the input thread.