direct3d11: use the indirect texture size
It can have padding for the decoder that is not known to avcodec and we don't
want to show it in vd->fmt to avoid adding a converter.

This padding needs to be taken in account when placing the picture in the display area.

We can get rid of the hackish BEFORE_UPDATE_RECTS/AFTER_UPDATE_RECTS macros.