Revert "picture: copy more than the visible lines"
This reverts commit 4594dedf117e40a1ba783ef55df995e12ea8e5fe.

This is the wrong way to do it. And breaks other things like magnify.
Either the visible area should account for the x/y offsets or there should be
and offset stored per plane.

We also need to decide in a picture copy copies only the visible area (and not
event the part in the x/y offset) or some of the decoder/crop padding too. The
current code leans towards the latter but still copies too much.