vout: remove the INIT control
It was posted exactly once as the first control and never again. Instead,
this handles the event implicitly when the video output thread starts.

The control queue processing flag must now be initially set, so that the
first WaitEmpty() actually waits for the video output thread to
initialize and get ready to process events.