bin: win32: Link vlc.exe and vlc-cache-gen.exe with -static
This prevents these executables from linking to a shared
libssp-0.dll if both a static and shared version is available
of this library.

Since these refer to the dynamically linked libvlc via libvlc.la,
it will still pick up the correct shared version of those, instead
of refusing to link to a dynamic library (as it does for any
library specified via -l<name>).

The same use of -Wc,-static is present in a few other Makefile.am
already.