variables: handle values when parsing bool
Currently, bool values can be set only without values, that is <varname> for
true or no-<varname> for false.

This causes an issue with the var_LocationParse function, that can't support
the "no-<varname>" bool syntax. This means that you can't set bool to false
when using this function.

To workaround this issue, this commit add support for value parsing, that is:
 - varname=0 for false, != 0 for true, or
 - varname=true for true, anything different for false.