skins2: fix fullscreen wrong size on Windows10 (high DPI)
When setting the scaling factor of Windows10 to something different
from the default 100%, a wrong fullscreen width/height was used.

At init, the skins2 engine is a non DPI-aware process. Since Windows10
expects to rescale in the background, reported sizes are altered.
Then, the skins2 engine launches the qt dialog provider, which makes the
process go from non DPI to DPI aware. As a result, Windows10 stops plans
to rescale, and provides the application with the real sizes.

The patch uses the latest valid sizes whatsoever.

This fixes trac #15349

TODO: As of today, the skins2 engine is seen as DPI aware,
(because of Qt) but nothing is implemented to behave accordingly.

(cherry picked from commit 028c89ad19d522976a1523e88f84102e667a7340)
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>