picture: align pictures on 64 bytes
Code relying on AVX-2 (like dav1d) cannot work without an alignment of 32 bytes
at least. Even avcodec has a requirement of 64 (likely to get at least 32 on all
planes). So it's probably time to upgrade.

picture_Setup ensures a horizontal alignment to 32 bytes and a vertical to 16
bytes. So for all buffers allocated by the core, we have a size multiple of 512
bytes.