es_out: also send es event when automatically unselected
Before, we always assumed the most common case: if you select an ES, you
unselect the old one automatically. This behavior is configurable (via
es_out_policy_e) and is not always true. Therefore, the es_out should always
send an event for consistency, in order to know if a track was really
unselected or not.