extras/tools: Add another libtool patch for building with lld on windows
On Windows, the maximum command line length is rather low; linking
libqt_plugin.la will reach this limit (but this seems to currently be
the only module linked which reaches the limit). In these cases, instead
of passing the list of object files to link on the command line, the list
is passed via a file. GCC/GNU ld supports both a response file and linker
scripts, while clang/lld only supports response files. This patch
reverses the priority between the two, by preferring response files.

This patch is included in the libtool package in msys2, which can be
considered some sort of active upstream.