core: playlist: expose "request" functions
When a user requests to insert, move or remove items, or to set the
current item, before the core playlist lock is successfully acquired,
another client may have changed the list.

Expose functions to apply the requested changes with potential conflicts
resolved. The actual changes applied are notified through the callbacks.

Signed-off-by: Thomas Guillem <thomas@gllm.fr>