input: handle "sub-fps" with a new control
We now change the slave demux rate according to the requested subtitle fps and
the main demuxer fps.

Example: Video is 24fps but the subtitle is 26fps. In that case, we want to
slow down the subtitle demuxer. Setting "sub-fps" to 26 will change the rate
of the subtitle demuxer to 0.92.