avcapture: Explicitly ask for consent before using the video device
On 10.14, the system asks the user for consent for using the video
device. But by default, it does not block the initialization, but
just outputs black frames before user consent is given.

This gives a bad experience for the user, also because the module
is not terminated if the user actually declines.

Instead, ask for consent explicitly, and block the open call to
wait for the response. On decline, Open directly returns with an
error, and the log contains the reason.