modules: common: use full name as MODULE_NAME
MODULE_NAME is only used for static builds:

Here is an example:

for codec/liba52_plugin_la-a52.lo
 MODULE_STRING=a52
 MODULE_NAME=codec_a52

for packetizer/a52.lo
 MODULE_STRING=a52
 MODULE_NAME=packetizer_a52

This patch will fix conflicts with the 2 vlc_entry__a52 symbols (that will now
be vlc_entry__codec_a52 and vlc_entry__packetize_a52).

On Android, we used objcopy to rename these symbols, but this tool is not
availaible on iOS toolchain.