macosx: Fix quitting of VLC over quit menu item
An event must be queued into the event loop so that the stop flag
can be executed and the main loop can be actually stopped.

If we we would have called the super implementation of this method
this would be not needed. But due to our custom deinit code in
Close, we do not call super, but instead do manual cleanup there.

closes #21369