snap: ship libaacs in the resulting snap
While this enables snapped VLC to play Blu-ray discs, the setup is a bit
more convoluted than the regular non-confined VLC, because
XDG_CONFIG_DIRS are redefined in the snap runtime.  This means AACS
config file must be put under $HOME/snap/vlc/current/.config/, and will
not be moved over when a new version is installed.