snap: make vulkan actually work inside a snap
Ubuntu 16.04 has an outdated libvulkan package which ships layers in
the -dev subpackage, so we have to add it to staged packages.  This can
be reverted when it's fixed.

Also introduced another part "fixup-vulkan-icd-paths".  The rationale to
add it is that icd ship with an absolute path to vulkan drivers, which
breaks under snap confinement;  with just a filename, vulkan will look
for those files under LD_LIBRARY_PATH.

(cherry picked from commit 58c57d1c4bdf17459f74dd70a3374f25cb38911e)