sftp: do not force TCP, SOCKS
libssh2 can accept any type of stream sockets - though that's largely
theoretical in this case - so there are no particular reasons to force
TCP here.

Regarding SOCKS, OpenSSH can be used to provide a server, but does not
generally act as a client. It seemed rather odd for VLC SFTP to use
SOCKS when normal command line SFTP does not.