gnutls: remove the run-time version check
This was wrong. If VLC is built with GnuTLS 3.5.0+, then it needs
GnuTLS 3.5.0+ at run-time, even if VLC can be built with GnuTLS 3.3.0.

In practice, this check is obviously pointless if GnuTLS is statically
linked. And if GnuTLS is dynamically linked, the packaging system shall
take care not to downgrade it to a run-time incompatible version (as
any other library). So just remove the check.

Instead, print the run-time and compile-time version numbers.