Qt(menu): fix RendererMenu for skins2
If a menu is given a parent, the menu will transparently be deleted once
the parent is deleted. If the menu is also kept in a static variable,
there is no easy means to reset said variable to NULL when the deletion
does occur. That leaves the likelihood to dereference an invalid pointer.

For the Qt interface, luckily the first parent object is the menu bar
whose lifetime is expected to span the entire vlc session, so the risk
is limited.

For the skins2 interface, on the contrary, the first parent object is a
popupmenu with a very short lifetime. So, it is bound to crash as soon
as a first popupmenu is displayed and closed.

This patch sets a NULL parent.