macosx: Fix correct reinitialization of most of our buttons
The button graphics were totally wrong if appearance changed in
the middle of the program. Also on startup, some graphics were wrong.
Factor out the code and use the same in all cases.

Also fix automatic appearance switching for volume slider, set
darkMode variable correctly, and fix appearance if setting was dark
mode and the OS mode is bright.