vout: libplacebo: add PL_COLOR_PRIM_DISPLAY_P3
This was added to the list of display primaries in libplacebo 0.5.0, but
we never added it to this list. No ifdef is needed because we already
require libplacebo >= 0.5.0 in our configure.ac.

We add it to the end of the list to avoid breaking existing vlcrc, even
though the order is less pretty this way.

Signed-off-by: Thomas Guillem <thomas@gllm.fr>