fb: remove "hardware accelerated" mode
This poorly named flag was meant to disable double buffering. I doubt
it ever worked. But if it did, that was before the video output rework
nine years ago.

Providing a pool of one picture would make the core copy manually
before the prepare callback. There are no differences to copying
manually in the display callback, except for exposing the picture too
early in the frame buffer. Just merge the two modes.