core: playlist: keep current item on expand
When an item is expanded (e.g. due to preparsing), it is replaced by its
(flatten) subitems.

The previous implementation first removed the item, then inserted its
subitems. As a consequence of the removal, the current index was reset
and the player was stopped.

Since we don't want to stop the player on item expansion, instead,
replace the item by its first subitem (if any), then insert the
remaining subitems.

Signed-off-by: Thomas Guillem <thomas@gllm.fr>