vout: drop last parameter of vout_display_PlacePicture()
Clipping the picture place computes incorrect values, which cannot be
used for correct rendering. Instead the place should never be clipped,
and the picture should be cropped if one or both coordinates are
negative, or one or both dimensions exceeds the display dimensions.

Fortunately, only one video output used the broken mode.