xcb/x11: clean up search for X11 visual
Pixmap formats are not indicative of what the screen can actually show.
For instance, 1 bit formats are typically used for pixel masks, but can
not be painted.

Instead look for a supported depth, then a supported visual for that
depth. Also check (in the previous commit helper) the RGB masks actually
corresponds to a VLC pixel format. We would not want to treat
R10G10B10A2 as R8B8G8A8.

This changeset effectively drops RGBA support for the time being.