vout: reset "pictures" at right time
This sends the pictures reset control to display, and updates the
display conversion filters at the needed time. That is to say after the
display format control requests are processed.

This fixes a race whereby a picture is allocated and rendered with the
old format after the reset request is sent but before it is processed.

Fixes #21580.