display: copy picture into pool if necessary
If the display supports the source format, the converter chain will be
empty, and the "conversion" will be a no-op. Yet, the picture might not
be allocated from the display pool, e.g. because the pool is slow or too
small compared to the DPB size.

Some filters will also do this when they modify the pictures in place.
However conversion filters should not be able to do that.

Copy the picture manually when this corner case happens.