vout: drop unnecessary pool check
Pictures are now systematically copied into the correct display pool if
necessary, so which pool the decoder allocates pictures from is
irrelevant.

Note though that this did not, and still does not, ensure that pictures
are in the expected format. A buggy decoder could and still can send
pictures in a wrong or old format.