display: do not copy opaque pictures
If the video format is opaque, the useful data is accessed through
picture_t.p_sys. As such, copying a picture into the display pool will
not work.

So do not copy pictures in that case, and instead assume that the
upstream converter or decoder agree with the display on the semantics
of picture_t.p_sys.