transcode: fix user filter codepath on transcoding video
Everything in chroma wise and size-wise are already in ok case without
this check, in this codepath we broke the i_chroma for encoder input fmt
and defined it to be encoder output i_chroma, which is usually the
codec. So filter chain ended up trying to find chroma-filtering between
I420 and h264, and failing.

Tested with
"#transcode{vfilter=canvas{width=640,height=640,aspect=16:9},vcodec=h264}:dummy"
and
"#transcode{width=640,height=640,vcodec=h264}:dummy"

Without this change, first case failed and second worked, with this code
removal, both cases worked and produced output file that was correct
resolution and size and playable.