spu: do nothing if input does not change
An input thread must be created before it is attached to the SPU,
and detached from the SPU before it is destroyed. As such, if the
already attached input pointer equals the newly attached input, we can
infer that it is actually the same input. That is then a no-op.