cea708: Fix support for rendering of multiple windows
In the existing code, p_region would always be equal to
p_spu_sys->region, since that is what it is initalized to and
nothing ever changes it.  Hence as the loop iterates, all the
windows will be update the first subpicture region (i.e.
corrupting that region such that it's properties are exclusively
that of the last window in the loop).  The result is only
the last window is ever rendered on-screen.

Add a variable to determine whether to use the original region
or whether to create a new region (so we end up with a different
subpicture region for each window).

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>