cea708: Fix TB scrolling
The decision when to truncate was being made based on the first
row in the array, but for TB scrolling we need to be checking
the last row.  Presumably this is a cut/paste error from the
above code block for BT scrolling.

Also, the for loop for the copy needs to include the first
entry, since i_firstrow gets incremented implicitly by the
truncate operation.

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>
(cherry picked from commit db96c022fa45a7618c93dcc8d6b81ebeb2d01e0f)