cea708: Fix when to scroll when in BT mode
When doing roll-up captions, the decision on whether to scroll
should not be based on the number of lines in the window, but
rather whether we were on the last line when we received the CR
command.

This problem manifests as the same line being overwritten
continuously rather than scrolling when we receive a sequence
such as the following:

DefineWindow with rc=1 (i.e. two lines)
SetPenLocation Row 1 (i.e. the second row)
Write caption
CarriageReturn
SetPenLocation Row 1
Write caption
CarriageReturn

In the above example, because CEA708_Window_RowCount() would
return 1, the row number would be bumped and a scroll would
not be performed.  Hence the subsequent SetPenLocation would
proceed to overwrite the current line.

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>