vout: ignore the display configuration when stopping
Each parameter generally fell in one of three category:
1) parameter that could not change:
   - window,
   - alignement,
   - pixel aspect ratio;
2) parameter whose value was ignored:
   - window size;
3) parameter whose value was actually saved from the old display
   and restored to the new display:
   - fill flag,
   - zoom,
   - viewpoint.

The last three are now correctly tracked in the vout configuration,
so they no longer need to be retrieved from the display being stopped.
(As a side benefit, the parameter can be tuned even if the vout is
inactive.)

This addresses the FIXME from c75fafe4ec348154a6e07c324dbe62c1c23a7149.