vout: separate stop event from reinit event
vout_Stop() only needs to deregister the mouse event handler; there is
essentially nothing in common with the vout_Request() REINIT control.

Eventually, vout_Stop() should probably do the ThreadStop() thing, but
that is not immediately possible - control requests may arrive before
the vout is restarted, and need to be processed.