vout: do not create new if reinit failed
Reinitializing a video output thread fails either due to unsupported
format is or a critical error such as OOM.

Either way, creating a new video output thread for the same format
right away is unlikely to lead to any good result. It is a waste of
time.

This patch should almost halve the delay in skipping unsupported video
format, when a vout is being reused.