vout: flush SPU channel on calling thread
There is a mixture of subpicture-related calls, with some running on
the calling thread, and others running on the video output thread
control loop.

The later may be delayed, depending on other video output thread
processing, sleeping and pending control requests. This leads to a
race condition, and potentially inconsistent ordering of calls on the
SPU instance.

This patch runs vout_FlushSubpictureChannel() directly on the calling
thread and adds a missing check for nonexistent/destroyed SPU.