lua: sd: Remove sd media browsing
Services discovery media were displayed via the old playlist, which allowed to
display their tree.

The new playlist is not a tree, but a list, so it can't be used to display
services discovery trees. This lua API is only used from cli.lua, and we
consider it's not worth it to implement this feature for now, so just remove
it.

Signed-off-by: Thomas Guillem <thomas@gllm.fr>