pulse: use pa_stream_begin_write()
Instead of using a free callback with a hackish way to get the block pointer.
Even if pa_stream_begin_write() is used to optimize the number of memcpy, this
commit won't change the total number of memcpy. The memcpy is just moved to the
Play() function.

pa_stream_begin_write() is available since 0.9.16. I guess that's why it was
not used since the beginning.