keystore: try memory keystore on the first try
This will speed up remote and protected accesses opening when the user already
entered a credential. Indeed, accessess generally call vlc_credential_get()
with NULL dialog title/fmt arguments first in order to get the password saved
from url/vlc-option but without triggering a dialog. Then if it is failing,
they call the same function with valid dialog arguments in order to trigger the
keystore and the dialog.

This commits will trigger the memory keystore from the first try. This will
prevent accesses to always try the default ("Guest") account if the user
already provided a valid credential.