messages: create one logger object per module
This creates a logger object only if/when a logger module is in use.
This avoids needing a logger object when there are no logger modules,
notably during early initialization. This also avoids having a logger
object carrying the name of a former logger module if logging is
changed during the lifetime of the VLC instance.

This requires wrapping, since the logger private data pointer is now
used to store the pointer to the VLC object. Note though that there
only ever is one or zero logger module in any given instance.