vout: actually stop the vout thread in vout_Stop()
This factors the vout thread stopping code from vout_Request() and
vout_Close() into vout_Stop(). vout_Request() now calls vout_Stop() when
it attempts to recycle an active vout (but not a stopped vout). Likewise
vout_Close() calls vout_Stop() if the vout being destroyed is active.

This makes no practical externally visible differences, but it is a
prerequisite to support creating a vout thread in stopped state.