aout: use original pts as play date
This will be required by the new vlc_clock_t. Indeed, in case of rate change,
only the filters are changed but not the aout. This means that aout->time_get()
timebase need to be the same than clock update points: the orignal pts given by
the decoder. In case the filters are outputting slightly more or less data
because of a ressampler imprecision, the delay will be corrected via a future
aout->time_get() call.