objects: use a single flat list
This internally replaces the objects tree with a flat list. This makes
browsing from parent to children slower, which is a bad idea and only
vlc_object_find_name() relies on.

It is far more common to go from child to parent, which is unaffected
and remains O(1). Although it is also not typically a good idea.