v4l2: put controls on the parent
This is not exactly great, and it (still) will not work properly if
there are more than one demux with controls on a single input.

But at least it does not rely on racy walking of the objects tree
and matching object name, nor racy usage vlc_object_t public members.