es_out: remove helpless decoder hold
If the decoder is valid for holding, then it is also valid for setting
a VLC object variable or adding a VLC object variable callback.
Otherwise, loading the decoder pointer or accessing the decoder is
undefined.

Either way, holding then releasing serves no purpose.

Note: input_GetEsObjects() should probably be removed in favor of the
new VBI input controls; but those are not currently exposed to input
managers.