avi: remove 10 ms sleeps
This originally was intended to leave the user a chance to stop the
input (assuming that vlc_object_t.b_die was atomic). But right now, it
leaves the demuxer infinitely. This does not even really solve the CPU
usage problem, as sleeping for 10 milliseconds might well be cause a
busy loop internally.

In practice, this nowadays relies on the VLC interrupt subsystem to
cause an I/O error if the user stops the (broken) input. That does not
depend on the demuxer sleeping to happen.

To completely address the problem, the demuxe callback should presumably
return after a few iterations, so that the input thread can pace and
eventually terminate cleanly.