opengl/display: fix option categorisation for gles2
note: the lack of cat/subcat on the gles2 module may have been deliberate,
due to the fact that gl and gles2 are separate plugins from the same
source and with this correction they introduce duplicate options when both
are installed on a system, which is not ideal, however there is a similar
issue with vout plugins and the fact that the options are guaranteed
identical means that it should not be a real issue. this also ensures that
the gles2 plugin will work correctly independent of the gl one (if such a
situation applies to anyone).

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>