mp4: fix help output bug
'cat hints' should only be used by the core (to define a set of
headings to break its large set of options up with, instead of the
option tree items).

incorrectly using a cat hint here instead of a section means that this
"hacks" sub-heading gets printed like a module heading, which is both
imperfect from a neatness point of view, but also and more importantly
potentially confusing.