vout:win32: handle the Always On Top of the non-parented window internally
When we receive our IDM_TOGGLE_ON_TOP we post WM_VLC_SET_TOP_STATE to our window.
No need to go through the vout_display_SendWindowState().

This is only used for the external window. In GUI/embedded mode the above/normal
state is handled by the GUI and we don't have to change menus in the chromeless
window.