macOS packaging: Set VLC_LIB_PATH for vlc-cache-gen call
Previously, both the installation dir and the plugins dir in VLC.app
were scanned. This lead to some objc warning due to having the same
objects in multiple currently loaded libraries.

This is just a cosmetic change to avoid the warnings, it should not
have any effect on the written plugin cache.