qt: provide base classes to render video output within the interface
VideoSurface provide a QQuickItem which can be embed within a QML scene to
represent the video node.

QVoutWindow implement a window where the video will be renderered.

VideoSurfaceProvider is the actual implementation of the VideoSurface which
will be selected at runtime. it will act as a the bridge between the QML
Videosurface and the QVoutWindow.

The new MainInterface methods return a null members during the qml transition.

Signed-off-by: Thomas Guillem <thomas@gllm.fr>