wall: remove aspect ratio parameter
The aspect ratio of window is fully determined by the aspect ratio of
the source video and the quotient of colums by rows. Adding an extra
parameter makes no sense, and does not work properly.

This fixes #17433, #213.

To select a custom ratio, the source video must be filtered with
cropadd. (In the future, we could also support crop in the splitter
display plugin.)