Set application display name and desktop entry
This ensures that the application can be properly identified.
For instance, a platform theme plugin may use those hints to identify the app
to the window manager, notification service, or system tray container.
Furthermore, on Wayland a desktop entry name is required in the the xdg-shell protocol.

Qt also automatically appends the application display name to the window title
on Windows and Linux removing the need for a custom window title function.

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>